+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * Enchancement: Improve EventProcessor nullability annotations (#1229).
 * Fix: Disable Gson HTML escaping
 * Enchancement: Add overloads for startTrasaction taking op and description (#1244)
+* Enchancement: Support @SentrySpan and @SentryTransaction on classes and interfaces. (#1243)
 
 # 4.1.0